все типа long

+++ src/SalesManages.java
@@ -1,13 +1,13 @@
 public class SalesManages {
-    protected int[] sales;
+    protected long[] sales;
 
-    public SalesManages(int[] sales) {
+    public SalesManages(long[] sales) {
         this.sales = sales;
     }
 
-    public int max() {
-        int max = -1;
-        for (int sale : sales) {
+    public long max() {
+        long max = -1;
+        for (long sale : sales) {
             if (sale > max) {
                 max = sale;
             }
